sort: use noted(NDFLT) in note handler
There are many more random notes flying around
here in Unix-land than there were on Plan 9.
For example, some shells implement "cat file | sort"
with cat as the child of sort, so that when cat exits,
sort gets a "sys: child" note.  noted(NDFLT) knows
which signals aren't really important and can be ignored,
and which need to kill the program.