This patch fixes unicode, unicode-be and unicode-le input, and unicode output, which have been broken due to the change of the size of Rune from two bytes to four bytes.
(Unicode-le and unicode-be output have not been
affected, since they work different, and produce
correct output as long as sizeof(Rune) >= 2.)

In all of the affected functions an extra array of N
elements of type u16int is used to temporarily store
the 16-bit unicode data.  As this increases stack
usage, perhaps a different, slightly more complicated
solution might be preferred, where only a single
`Rune buf[N]' is used for both temporary storage and
the result.

R=rsc
CC=codebot
http://codereview.appspot.com/1574041