INSTALL: build arm64 binaries on arm64 Macs
If there is any x86 binary in the chain of parent processes from
the current one, then uname -m prints x86_64 and clang
defaults to building x86_64 binaries.

Detect arm64 using the kernel version and force building
an arm64 toolchain instead of perpetuating x86.

This is particularly important when the user shell is rc,
which has been built for x86, because then rebuilding
under rc continues to use x86.